fix: split up owlbot-bootrstapper into 3 sections
