+++ packages/owlbot-bootstrapper/app/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "googleapis-bootstrapper-app",
+  "version": "1.0.0",
+  "description": "Responds to GH webhook events to bootstrap the initial files on new APIs to generate libraries",
+  "private": true,
+  "author": "Google LLC.",
+  "license": "Apache-2.0",
+  "repository": "https://github.com/googleapis/repo-automation-bots.git",
+  "homepage": "https://github.com/googleapis/repo-automation-bots",
+  "bugs": "https://github.com/googleapis/repo-automation-bots/issues",
+  "main": "build/src/app.js",
+  "keywords": [
+    "probot",
+    "github",
+    "probot-app"
+  ],
+  "files": [
+    "build/src"
+  ],
+  "scripts": {
+    "compile": "tsc -p .",
+    "start": "probot run ./build/src/owlbot-bootstrapper.js",
+    "pretest": "npm run compile",
+    "test": "cross-env LOG_LEVEL=fatal c8 mocha --exit build/test",
+    "fix": "gts fix",
+    "lint": "gts check"
+  },
+  "dependencies": {
+    "probot": "^12.2.2"
+  },
+  "devDependencies": {
+    "@types/mocha": "^8.0.0",
+    "c8": "^7.3.5",
+    "cross-env": "^7.0.0",
+    "gts": "^3.0.0",
+    "mocha": "^8.0.0",
+    "typescript": "~4.4.3"
+  },
+  "engines": {
+    "node": ">= 12.18.2"
+  }
+}
