+++ packages/owlbot-bootstrapper/app/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "gts/tsconfig-google",
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "rootDir": ".",
+    "outDir": "build"
+  },
+  "include": [
+    "*.ts",
+    "test/*.ts",
+  ]
+}
