+++ packages/owlbot-bootstrapper/cli/.c8rc.json
@@ -1,0 +1,3 @@
+{
+  "exclude": [".mocharc.js", "build/test"]
+}
