+++ packages/owlbot-bootstrapper/cli/.dockerignore
