+++ packages/owlbot-bootstrapper/cli/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./node_modules/gts"
+}
