+++ packages/owlbot-bootstrapper/common-container/.c8rc.json
@@ -1,0 +1,3 @@
+{
+  "exclude": [".mocharc.js", "build/test"]
+}
