+++ packages/owlbot-bootstrapper/common-container/.dockerignore
@@ -1,0 +1,7 @@
+build/
+coverage/
+node_modules/
+.env
+*.pem
+Dockerfile
+.dockerignore
