+++ packages/owlbot-bootstrapper/common-container/.eslintignore
@@ -1,0 +1,4 @@
+**/node_modules
+**/coverage
+test/fixtures
+build/
