+++ packages/owlbot-bootstrapper/common-container/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./node_modules/gts"
+}
