+++ packages/owlbot-bootstrapper/common-container/.gitignore
@@ -1,0 +1,3 @@
+node_modules
+build
+coverage
