+++ packages/owlbot-bootstrapper/common-container/.prettierignore
@@ -1,0 +1,6 @@
+**/node_modules
+**/coverage
+test/fixtures
+build/
+docs/
+protos/
