+++ packages/owlbot-bootstrapper/common-container/commands/pre-process.ts
@@ -13,21 +13,27 @@
 // limitations under the License.
 
 import yargs from 'yargs';
-import {CliArgs} from '../../common-container/interfaces';
-import {postProcess} from '../../common-container/post-process';
+import {preProcess} from '../pre-process';
 
-export const postProcessCommand: yargs.CommandModule<{}, CliArgs> = {
-  command: 'post-process',
-  describe: 'Run post-process steps for owlbot-bootstrapper',
+interface CliArgs {
+  projectId: string;
+  triggerId?: string;
+  apiId: string;
+  repoToClone?: string;
+  language: string;
+  installationId: string;
+  container: string;
+  languageContainer?: string;
+  buildId?: string;
+}
+
+export const preProcessCommand: yargs.CommandModule<{}, CliArgs> = {
+  command: 'pre-process',
+  describe: 'Run pre-process steps for owlbot-bootstrapper',
   builder(yargs) {
     return yargs
       .option('projectId', {
         describe: 'project ID which contains the build file',
-        type: 'string',
-        demand: true,
-      })
-      .option('triggerId', {
-        describe: 'trigger of build to run',
         type: 'string',
         demand: true,
       })
@@ -64,9 +70,10 @@ export const postProcessCommand: yargs.CommandModule<{}, CliArgs> = {
       .option('buildId', {
         describe: 'build id of given build',
         type: 'string',
+        demand: false,
       });
   },
   async handler(argv: CliArgs) {
-    postProcess(argv);
+    preProcess(argv);
   },
 };
