+++ packages/owlbot-bootstrapper/common-container/package.json
@@ -19,8 +19,6 @@
   ],
   "scripts": {
     "compile": "tsc -p .",
-    "start": "probot run ./build/src/googleapis-bootstrapper.js",
-    "start:local": "node ./build/common-container/common-container.js",
     "pretest": "npm run compile",
     "test": "cross-env LOG_LEVEL=fatal c8 mocha --exit build/test",
     "fix": "gts fix",
