+++ packages/owlbot-bootstrapper/common-container/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "gts/tsconfig-google",
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "rootDir": ".",
+    "outDir": "build"
+  },
+  "include": [
+    "*.ts",
+    "test/*.ts",
+    "commands/*.ts",
+  ]
+}
