--- packages/owlbot-bootstrapper/test/fixtures/fake-private-key.pem
@@ -1,1 +1,0 @@
------BEGIN RSA PRIVATE KEY-----\nMIICXAIBAAKBgQCSF/xZBi6SFEak4Frj+PyUer3W06nQifm0GiLJqayMItRTxD+Y\nvv1PI1l8rhFNsnmKu1GPPq2itEZRwohaBL7xddVRRnrcn3S28/qswSnFKnje1Mg5\nVaMTk/axUIbRJ48MvX3BV3iDg7UgX2+LFw5Dd1iimaXcZ8TTNzeXrtrMtQIDAQAB\nAoGAOeMnKl84tMkz9MCxQwO1kV/ROP5Fm0wQlTyF7EI/DIgAefCZAn0rAeCP5MdU\nM+U0I7R1bLz5MC8ebecXABRBU/JmFUDFfaNz0XuOT+C2ySdHTjqHuuqXoFeSJvpW\npxyG7d87Ka5FM+9FfrWbvQBCk8c/uoFdULtXV36lrX0V70ECQQDJ6xoaw3nTeB4F\nCQgg8ckQXUaVnaLEiZX6V/P0jIBmL/dCt9FpnT+I/JGtlpNJipWt7enl/IOd5VMT\nAwGDq4gvAkEAuTkqAupZZwYDQFYBijppRaMwwY1uy5fiZiy2nIOHTKsqjhXR3e+/\nwV78NZvmVXvo/1Os3G/pWBF5TMVD9nfcWwJAGTYwyTy1HnF2Fhzdlfd6qn9rC81m\nAHJJdru6q0NL2PaOQhdfn3dR8pHOwsT1WQ5jyBbXyK6x5Go46DvM2Zr2YwJBAJOY\nQKjoE2ur9+VJY5rBZuet3YkuaY5krmzazn0zthna/2MPcheXF0dAqFsTqtyNi6IC\nDGnv55JdAjqxqrF426cCQEKx2+enOsBKyLxeHfaBHv34em96VCqUVCFyvAXivKeb\nCCdgkD3tRfBCB99XbxtUrFSmbZ/l8zPQ3SGYDjoi4Y8=\n-----END RSA PRIVATE KEY-----
