--- packages/owlbot-bootstrapper/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "gts/tsconfig-google",
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "rootDir": ".",
-    "outDir": "build"
-  },
-  "include": [
-    "src/*.ts",
-    "test/*.ts",
-    "cli/*.ts",
-    "common-container/*.ts", "test/common-container/utils.test.ts", "test/common-container/split-repo.test.ts", "test/app/run-trigger.test.ts", "test/common-container/mono-repo.test.ts", "test/common-container/github-authenticator.test.ts", "test/common-container/common-container.test.ts" ]
-}
